backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
httpx==0.28.1
python-jose[cryptography]==3.4.0
python-multipart==0.0.20
pydantic-settings==2.9.1
SQLAlchemy==2.0.29\npsycopg2-binary==2.9.9
